requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.12,<3.14
drf-spectacular>=0.22.0,<0.23.0
celery>=5.2,<5.3
redis>=4.2,<4.4
dj-database-url>=0.5.0,<0.6.0
gunicorn>=20.1.0,<21.0
whitenoise>=6.0.0,<7.0.0
openai>=1.0.0,<2.0.0
pydantic>=1.9.0,<2.0.0
python-dotenv>=0.19.0,<1.0.0
pytest>=6.2,<7.0
pytest-django>=4.5,<4.6
django-filter
